+++ src/main/java/com/cnam/nfa036projet/controller/StockController.java
@@ -3,12 +3,8 @@ package com.cnam.nfa036projet.controller;
 
 import com.cnam.nfa036projet.form.*;
 import com.cnam.nfa036projet.model.*;
-import com.cnam.nfa036projet.repository.ProduitRepository;
-import com.cnam.nfa036projet.repository.StatutRepository;
-import com.cnam.nfa036projet.repository.StockHistoriqueRepository;
-import com.cnam.nfa036projet.repository.StockRepository;
-import com.cnam.nfa036projet.service.StockService;
-import com.cnam.nfa036projet.service.UtilisateurDetailsService;
+import com.cnam.nfa036projet.repository.*;
+import com.cnam.nfa036projet.service.*;
 import com.lowagie.text.DocumentException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -28,8 +24,10 @@ import java.io.*;
 import java.nio.file.FileSystems;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.time.LocalTime;
 import java.time.temporal.ChronoUnit;
 import java.util.*;
+
 
 @Controller
 public class StockController {
@@ -48,142 +46,23 @@ public class StockController {
     @Autowired
     private StockHistoriqueRepository stockHistoriqueRepository ;
 
-    private StockService stockService = new StockService() ;
+    @Autowired
+    private StockService stockService ;
+
+    @Autowired
+    private StockHistoriqueService stockHistoriqueService ;
+
+    @Autowired
+    private UtilisateurService userService ;
+
 
     public Statut findByNomStatut(String nomStatut){
         Statut statut = statutRepository.findByNomStatut(nomStatut);
         return statut ;
     }
 
-    /*
-     *Méthode pour récupérer le nom + prénom de l'Utilisateur connecté
-     *
-     */
-/*
-    public String getNomUser(){
-        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-
-        if (principal instanceof UtilisateurDetails) {
-            return ((UtilisateurDetails)principal).getNomPrenom();
-        } else {
-            return principal.toString();
-        }
-    }
-*/
-    //Méthode d'envoi de la Liste de Produits en Stock, étant appelée après chaque opération
-
-    public List<StockForm> listeStock() {
-        List<Stock> stocks = stockRepository.findAll();
-        List<StockForm> stockList = new ArrayList<>();
-        for (Stock stock:stocks) {
-
-                //MISE A JOUR DU STATUT SI DLC COURTE = INFERIEUR A 24 HEURES + SAUVEGARDE NV STATUT EN BASE + SAUVEGARDE HISTORIQUE
-                LocalDateTime dateEntree = stock.getDateEntree().truncatedTo(ChronoUnit.SECONDS);
-                LocalDateTime dlc = dateEntree.plusDays(stock.getProduit().getDureeConservation());
-                dlc.truncatedTo(ChronoUnit.SECONDS);
-                long daysBetween = LocalDateTime.now().until(dlc, ChronoUnit.HOURS);
-                if (stock.getStatut().getNomStatut().equals("En Stock") && daysBetween <= 24) {
-                    Statut statut = findByNomStatut("A Contrôler");
-                    stock.changeStatut(statut);
-                    stockRepository.save(stock);
-
-                    // AJOUT D'UN NOUVEL HISTORIQUE
-                    StockHistorique historique = new StockHistorique();
-                    historique.setDateMouvementStock(LocalDateTime.now());
-                    historique.setProduit(stock.getProduit().getNomProduit());
-                    historique.setIdProduit(stock.getId());
-                    historique.setUtilisateur(UtilisateurDetailsService.getNomUser());
-                    historique.setStatut(statut.getNomStatut());
-                    historique.setCategorie(stock.getProduit().getCategorie().getNomCategorie());
-                    stockHistoriqueRepository.save(historique);
-                }
-
-                //On renseigne stockForm
-                long id = stock.getId();
-                String nomProduit = stock.getProduit().getNomProduit();
-                String categorie = stock.getProduit().getCategorie().getNomCategorie();
-                String statut = stock.getStatut().getNomStatut();
-                int nbStatut = 0 ;
-                if (daysBetween > 24) {
-                    nbStatut = 1;
-                } else if (daysBetween >= 0) {
-                    nbStatut = 2;
-                } else nbStatut = 3;
-
-                StockForm aStock = new StockForm(id, nomProduit, categorie, dateEntree, dlc, statut, nbStatut);
-                stockList.add(aStock);
-        }
-        return stockList ;
-    }
-
-    //HISTORIQUE DES STOCKS SELON UNE DATE
-
-    public List<HistoriqueForm> historiqueStock(LocalDate date) {
-        List<StockHistorique> stocks = stockHistoriqueRepository.findAll();
-        List<HistoriqueForm> stockList = new ArrayList<>();
-        for (StockHistorique stock:stocks) {
-            //renseignement de la liste selon la date
-            if( stock.getDateMouvementStock().toLocalDate().isEqual(date) )
-            {
-                long idProduit = stock.getIdProduit();
-                String nomProduit = stock.getProduit();
-                String categorie = stock.getCategorie();
-                String statut = stock.getStatut();
-                String utilisateur = stock.getUtilisateur();
-                LocalDateTime dateMouvement = stock.getDateMouvementStock().truncatedTo(ChronoUnit.SECONDS);
-
-                HistoriqueForm aStock = new HistoriqueForm(idProduit, nomProduit, categorie, statut, dateMouvement, utilisateur);
-                stockList.add(aStock);
-            }
-        }
-        //Tri de la liste selon ID stock
-        //A FAIRE
-        //
-
-        return stockList ;
-    }
-
-    //HISTORIQUE DES STOCKS SELON UNE ID PRODUIT
-
-    public List<HistoriqueForm> historiqueStockById(long id) {
-        List<StockHistorique> stocks = stockHistoriqueRepository.findAll();
-        List<HistoriqueForm> stockList = new ArrayList<>();
-        for (StockHistorique stock:stocks) {
-            //renseignement de la liste selon l'ID produit
-            if( stock.getIdProduit() == id )
-            {
-                long idProduit = stock.getIdProduit();
-                String nomProduit = stock.getProduit();
-                String categorie = stock.getCategorie();
-                String statut = stock.getStatut();
-                String utilisateur = stock.getUtilisateur();
-                LocalDateTime dateMouvement = stock.getDateMouvementStock().truncatedTo(ChronoUnit.SECONDS);
-
-                HistoriqueForm aStock = new HistoriqueForm(idProduit, nomProduit, categorie, statut, dateMouvement, utilisateur);
-                stockList.add(aStock);
-            }
-        }
-        //Tri de la liste selon DATE
-        //A FAIRE
-        //
-
-        return stockList ;
-    }
 
 
-    /**
-     * PAGE INDEX AVEC TABLEAU DE BORD
-     *
-     * READ
-     *
-     *
-     */
-
-    @RequestMapping(value = {"/", "/index"}, method = RequestMethod.GET)
-    public String home(Model model) {
-        model.addAttribute("stockList", listeStock());
-        return "/index";
-    }
 
     /**
      * LISTE DES PRODUITS EN MOUVEMENT DE STOCK
@@ -195,7 +74,7 @@ public class StockController {
 
     @RequestMapping(value = { "/readStock"}, method = RequestMethod.GET)
     public String readStock(Model model) {
-        model.addAttribute("stockList", listeStock());
+        model.addAttribute("stockList", stockService.listeStock());
         DateStockForm dateStock = new DateStockForm();
         model.addAttribute("dateStock",dateStock);
         return "/Stock/readStock";
@@ -214,7 +93,7 @@ public class StockController {
         if(bindingResult.hasErrors()){
             return readStock(model);
         } else {
-            model.addAttribute("historiqueList", historiqueStock(LocalDate.parse(dateHistorique.getDate())));
+            model.addAttribute("historiqueList", stockHistoriqueService.historiqueStock(LocalDate.parse(dateHistorique.getDate())));
             DateStockForm dateStock = new DateStockForm();
             model.addAttribute("dateStock",dateStock);
             return "/Stock/readHistoriqueStock";
@@ -235,7 +114,7 @@ public class StockController {
         if(bindingResult.hasErrors()){
             return readStock(model);
         } else {
-            model.addAttribute("historiqueList", historiqueStockById(dateHistorique.getId()));
+            model.addAttribute("historiqueList", stockHistoriqueService.historiqueStockById(dateHistorique.getId()));
             DateStockForm dateStock = new DateStockForm();
             model.addAttribute("dateStock",dateStock);
             return "/Stock/readHistoriqueStock";
@@ -284,7 +163,7 @@ public class StockController {
             Statut statut = findByNomStatut("En Stock");
             statut.addStock(stock);
             historique.setStatut(statut.getNomStatut());
-            historique.setUtilisateur(UtilisateurDetailsService.getNomUser());
+            historique.setUtilisateur(userService.getNomUser());
 
             stockRepository.save(stock);
             historique.setIdProduit(stock.getId());
@@ -295,8 +174,8 @@ public class StockController {
             LocalDateTime dlc = dateEntree.plusDays(stock.getProduit().getDureeConservation());
             dlc.truncatedTo(ChronoUnit.SECONDS);
             EtiquetteForm etiquette = new EtiquetteForm(stock.getId(), stock.getProduit().getNomProduit(), dateEntree, dlc );
-            String templateHtml = parseEtiquetteTemplate(etiquette) ;
-            generatePdfFromHtml (templateHtml) ;
+            String templateHtml = EtiquetteService.parseEtiquetteTemplate(etiquette) ;
+            EtiquetteService.generatePdfFromHtml (templateHtml) ;
         }
         //Retour à la liste en Stock
         return readStock(model);
@@ -344,7 +223,7 @@ public class StockController {
         historique.setDateMouvementStock(LocalDateTime.now());
         historique.setProduit(stock.getProduit().getNomProduit());
         historique.setIdProduit(stock.getId());
-        historique.setUtilisateur(UtilisateurDetailsService.getNomUser());
+        historique.setUtilisateur(userService.getNomUser());
         historique.setStatut(statut.getNomStatut());
         historique.setCategorie(stock.getProduit().getCategorie().getNomCategorie());
         stockHistoriqueRepository.save(historique);
@@ -352,53 +231,5 @@ public class StockController {
         return readStock(model);
     }
 
-    //Méthode de parsing de l'étiquette
-    private String parseEtiquetteTemplate(EtiquetteForm etiquette) throws UnsupportedEncodingException {
-        ClassLoaderTemplateResolver templateResolver = new ClassLoaderTemplateResolver();
-        templateResolver.setSuffix(".html");
-        templateResolver.setTemplateMode(TemplateMode.HTML);
-        templateResolver.setCharacterEncoding("UTF-8");
-
-        TemplateEngine templateEngine = new TemplateEngine();
-        templateEngine.setTemplateResolver(templateResolver);
-
-        Context context = new Context();
-        context.setVariable("Etiquette", etiquette);
-
-        String htmlContent = templateEngine.process("/templates/Stock/etiquetteStock",context);
-        return convertToXhtml(htmlContent);
-    }
-
-    //Méthode de génération de l'étiquette en PDF
-    public void generatePdfFromHtml(String html) throws DocumentException, IOException {
-        /*File file = File.createTempFile("etiqu", ".pdf");
-        OutputStream outputStream = new FileOutputStream(file) ;
-*/
-        String baseUrl = FileSystems
-                .getDefault()
-                .getPath("src", "main", "resources")
-                .toUri()
-                .toURL()
-                .toString();
-        ITextRenderer renderer = new ITextRenderer();
-        renderer.setDocumentFromString(html, baseUrl);
-        renderer.layout();
-
-        OutputStream outputStream = new FileOutputStream("etiquette.pdf");
-        renderer.createPDF(outputStream);
-
-        outputStream.close();
-    }
-
-    private String convertToXhtml(String html) throws UnsupportedEncodingException {
-        Tidy tidy = new Tidy();
-        tidy.setInputEncoding("UTF-8");
-        tidy.setOutputEncoding("UTF-8");
-        tidy.setXHTML(true);
-        ByteArrayInputStream inputStream = new ByteArrayInputStream(html.getBytes("UTF-8"));
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream();
-        tidy.parseDOM(inputStream, outputStream);
-        return outputStream.toString("UTF-8");
-    }
 
 }
