+++ src/main/java/com/cnam/nfa036projet/form/TempDuJour.java
@@ -1,0 +1,38 @@
+package com.cnam.nfa036projet.form;
+
+public class TempDuJour {
+
+    private String nomFrigo ;
+
+    private float tempMatin ;
+
+    private float tempAprem ;
+
+    public TempDuJour() {
+    }
+
+    public String getNomFrigo() {
+        return nomFrigo;
+    }
+
+    public void setNomFrigo(String nomFrigo) {
+        this.nomFrigo = nomFrigo;
+    }
+
+    public float getTempMatin() {
+        return tempMatin;
+    }
+
+    public void setTempMatin(float tempMatin) {
+        this.tempMatin = tempMatin;
+    }
+
+    public float getTempAprem() {
+        return tempAprem;
+    }
+
+    public void setTempAprem(float tempAprem) {
+        this.tempAprem = tempAprem;
+    }
+}
+
