+++ src/main/resources/static/styles/css/style.css
@@ -6,6 +6,7 @@ h2 {
     color:#5C8B26;
 }
 
+
 table {
     border-collapse: collapse;
 }
@@ -18,24 +19,112 @@ table th, table td {
     color:red;
 }
 
-header {
-    height: 130px;
-    background-color: #265C8B;
-    color: #E7E7E7;
-    font-weight: bold;
-}
-
-footer {
-    height: 120px;
-    background-color: #265C8B;
-    color: white;
-}
 
 .gras{
     font-weight: bolder;
     font-size: 1.1em;
 }
 
-#floatright {
+.floatright {
     float: right;
+}
+
+.aligncenter {
+    text-align: center ;
+}
+
+/* ------------------    BOUTONS INPUT + RESET DANS LE BODY    -----------------  */
+
+input[type=submit], input[type=reset] {
+    padding:5px 15px;
+    background:#265C8B;
+    color: white;
+    border:0 none;
+    cursor:pointer;
+    -webkit-border-radius: 5px;
+    border-radius: 5px;
+}
+
+input[type=submit]:hover, input[type=reset]:hover  {
+    padding:5px 15px;
+    background: #8B265C ;
+    color: white;
+    border:0 none;
+    cursor:pointer;
+    -webkit-border-radius: 5px;
+    border-radius: 5px;
+    font-size: 1.1em;
+}
+
+
+/* ------------------    ICONES DANS LES TABLEAUX    -----------------  */
+
+td a i {
+    color:#5C8B26;
+}
+
+td a i:hover{
+    color:#8B265C;
+    font-size:1.2em ;
+}
+
+/* ------------------    HEADER    -----------------  */
+
+header {
+    height: 130px;
+    background-color: #265C8B;
+    color: #E7E7E7;
+    font-weight: bold;
+    padding: 5px;
+    padding-left: 10px;
+}
+
+header img {
+    padding-right: 10px;
+    padding-top: 10px;
+}
+
+
+/* ------------------    FOOTER    -----------------  */
+
+footer {
+    height: 45px;
+    background-color: #265C8B;
+    color: white;
+    padding-left: 10px;
+    padding-top: 5px;
+}
+
+footer a:link, footer a:visited  {
+    color: white;
+    padding: 5px 15px;
+    text-align: center;
+    text-decoration: none;
+    display: inline-block;
+}
+
+footer a:hover, footer a:active {
+    background-color: white;
+    color: black;
+    border-radius: 5px;
+}
+
+footer input[type=submit] {
+    padding:5px 15px;
+    background:#265C8B;
+    color: white;
+    border:0 none;
+    cursor:pointer;
+    -webkit-border-radius: 5px;
+    border-radius: 5px;
+}
+
+footer input[type=submit]:hover {
+    padding:5px 15px;
+    background:white;
+    color: black;
+    border:0 none;
+    cursor:pointer;
+    -webkit-border-radius: 5px;
+    border-radius: 5px;
 }
