+++ src/main/resources/templates/Fragments/header.html
@@ -1,0 +1,34 @@
+<!DOCTYPE html>
+<html lang="fr" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="author" content="Guillaume Rabiller" />
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="description" content="header des pages HTML">
+
+    <!-- CSS + Bootstrap CSS -->
+    <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
+    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
+
+    <title>header des pages HTML</title>
+</head>
+
+<body class="container">
+
+
+
+        <header role="banner">
+            <img class="floatright" th:src="@{/pics/logo.png}" alt="logo logiciel" width="100" height="100">
+            <p class="gras aligncenter">TPR - Traçage Produits Restaurant</p>
+            <p class="d-none d-md-none d-lg-block"> Logiciel de gestion de stock restaurant  </p>
+            <p th:inline="text">Bienvenue [[${#httpServletRequest.remoteUser}]] !</p>
+
+
+        </header>
+
+
+
+
+</body>
+</html>
