+++ src/main/resources/templates/Produit/createProduit.html
@@ -9,6 +9,7 @@
     <!-- CSS + Bootstrap CSS -->
     <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
+    <script src="https://kit.fontawesome.com/610665bae1.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
 
     <title>Formulaire création produit</title>
@@ -19,21 +20,19 @@
 
     <!-- ------------------------    HEADER    ---------------------------- -->
 
+    <div th:replace="Fragments/header :: header"></div>
+
+    <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
+
+
     <div class="row">
-
-        <header class="col-12" role="banner">
-            <img id="floatright" th:src="@{/pics/logo.png}" alt="logo logiciel" width="100" height="100">
-            <p class="gras">TPR - Traçage Produits Restaurant</p>
-            <p class="d-none d-md-none d-lg-block"> Logiciel de gestion de stock restaurant  </p>
-            <p th:inline="text">Bienvenue [[${#httpServletRequest.remoteUser}]]!</p>
-
-
-        </header>
-
+        <a class="btn btn-outline-secondary" th:href="@{/index}" role="button"><i class="fas fa-home"></i> Retour au Tableau de Bord</a>
     </div>
 
 
-    <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
+    <br>
+    <br>
+
 
    <div class="row">
         <h1>Création d'un Produit</h1>
@@ -45,13 +44,13 @@
                 <label>Nom du produit</label>
                 <input type="text" th:field="*{nomProduit}" />
                 <span class="error" th:if="${#fields.hasErrors('nomProduit')}">Erreur sur le nom</span>
-                <br/>
+                <br>
             </div>
             <div class="form-group">
                 <label>Durée de conservation</label>
                 <input type="number" th:field="*{dureeConservation}" />
                 <span class="error" th:if="${#fields.hasErrors('dureeConservation')}" th:errors="*{dureeConservation}"></span>
-                <br/>
+                <br>
             </div>
             <div class="form-group">
                 <label>Catégorie</label>
@@ -60,31 +59,23 @@
                                 th:value="${categorieList.id}"
                                 th:text="${categorieList.nomCategorie}"/>
                 </select>
-                <br/>
+                <br>
             </div>
                 <input type="submit" value="Submit" />
                 <input type="reset" value="Reset" />
         </form>
    </div>
 
+    <br>
+    <br>
+
+
 
     <!-- ------------------------    FOOTER    ---------------------------- -->
 
-    <div class="row">
-        <footer class="col-12" role="contentinfo">
-            <div class="row">
-                <div class="col-3">
-                    <a href="#" title="about">A propos</a>
-                </div>
-                <div class="col-3">
-                    <a href="mailto:g.rabiller44@gmail.com" title="lien email" target="_blank">Service Client </a>
-                </div>
-                <form th:action="@{/logout}" method="post">
-                    <input type="submit" value="Sign Out"/>
-                </form>
-            </div>
-        </footer>
-    </div>
+    <div th:replace="Fragments/footer :: footer"></div>
+
+
 
 </body>
 </html>
