+++ src/main/resources/templates/Stock/readStock.html
@@ -9,6 +9,7 @@
     <!-- CSS + Bootstrap CSS -->
     <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
+    <script src="https://kit.fontawesome.com/610665bae1.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
 
     <title>Etat du stock</title>
@@ -18,48 +19,20 @@
 
     <!-- ------------------------    HEADER    ---------------------------- -->
 
-    <div class="row">
-
-        <header class="col-12" role="banner">
-            <img id="floatright" th:src="@{/pics/logo.png}" alt="logo logiciel" width="100" height="100">
-            <p class="gras">TPR - Traçage Produits Restaurant</p>
-            <p class="d-none d-md-none d-lg-block"> Logiciel de gestion de stock restaurant  </p>
-            <p th:inline="text">Bienvenue [[${#httpServletRequest.remoteUser}]]!</p>
-
-
-        </header>
-
-    </div>
+    <div th:replace="Fragments/header :: header"></div>
 
     <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
 
     <div class="row">
-    <a class="btn btn-outline-secondary" th:href="@{/index}" role="button">Retour au Tableau de Bord</a>
+    <a class="btn btn-outline-secondary" th:href="@{/index}" role="button"><i class="fas fa-home"></i>  Retour au Tableau de Bord</a>
     </div>
 
+
+    <br>
+    <br>
+
     <h1>Etat du stock :</h1>
-    <div class="row">
-        <form sec:authorize="hasRole('ADMIN')" th:action="@{/readHistoriqueStock}"
-              th:object="${dateStock}" method="POST">
-            <div class="form-group">
-                <label>Historique du Stock</label>
-                <input type="date" th:field="*{date}" />
-                <span class="error" th:if="${#fields.hasErrors('date')}" th:errors="*{date}"></span>
-                <br/>
-            </div>
-            <input type="submit" value="Submit" />
-        </form>
-        <form sec:authorize="hasRole('ADMIN')" th:action="@{/readHistoriqueStockById}"
-              th:object="${dateStock}" method="POST">
-            <div class="form-group">
-                <label>Historique par ID</label>
-                <input type="text" th:field="*{id}" />
-                <span class="error" th:if="${#fields.hasErrors('id')}" th:errors="*{id}"></span>
-                <br/>
-            </div>
-            <input type="submit" value="Submit" />
-        </form>
-    </div>
+
     <div class="row">
         <table border="1" class="table table-striped table-hover">
             <tr>
@@ -80,38 +53,61 @@
                 <td th:text="${stock.statut}"></td>
                 <td><a th:href="@{/updateStock/{id}(id=${stock.id})}">Edit</a></td>
             </tr>
-            <tr th:each="stock:${stockList}" th:if="${stock.colorStatut} != 1 " th:style="${stock.colorStatut} == 2 ? 'background: orange' : 'background: red'  " >
+            <tr th:each="stock:${stockList}" th:if="${stock.colorStatut} != 1 " >
                 <td th:text="${stock.id}"></td>
                 <td th:text="${stock.nomProduit}"></td>
                 <td th:text="${stock.categorie}"></td>
                 <td th:text="${stock.dateEntree}"></td>
                 <td th:text="${stock.dlc}"></td>
-                <td th:text="${stock.statut}"></td>
+                <td th:text="${stock.statut}" th:style="${stock.colorStatut} == 2 ? 'background: #CD710B' : 'background: #A73000'  " ></td>
                 <td><a th:href="@{/updateStock/{id}(id=${stock.id})}">Edit</a></td>
             </tr>
         </table>
     </div>
 
 
-    <a class="btn btn-info" th:href="@{/createStock}">Ajout d'un produit en stock</a>
+    <a class="btn btn-outline-primary" th:href="@{/createStock}">Ajout d'un produit en stock</a>
+
+    <br>
+    <br>
+
+    <div class="row">
+        <div class="col">
+            <form sec:authorize="hasRole('ADMIN')" th:action="@{/readHistoriqueStock}"
+                  th:object="${dateStock}" method="POST">
+                <div class="form-group">
+                    <label>Historique du Stock</label>
+                    <input type="date" th:field="*{date}" />
+                    <span class="error" th:if="${#fields.hasErrors('date')}" th:errors="*{date}"></span>
+                    <br>
+                </div>
+                <input type="submit" value="Submit" />
+            </form>
+        </div>
+        <div class="col">
+            <form sec:authorize="hasRole('ADMIN')" th:action="@{/readHistoriqueStockById}"
+                  th:object="${dateStock}" method="POST">
+                <div class="form-group">
+                    <label>Historique par ID</label>
+                    <input type="text" th:field="*{id}" />
+                    <span class="error" th:if="${#fields.hasErrors('id')}" th:errors="*{id}"></span>
+                    <br>
+                </div>
+                <input type="submit" value="Submit" />
+            </form>
+        </div>
+    </div>
+
+
+    <br>
+    <br>
+
 
     <!-- ------------------------    FOOTER    ---------------------------- -->
 
-    <div class="row">
-        <footer class="col-12" role="contentinfo">
-            <div class="row">
-                <div class="col-3">
-                    <a href="#" title="about">A propos</a>
-                </div>
-                <div class="col-3">
-                    <a href="mailto:g.rabiller44@gmail.com" title="lien email" target="_blank">Service Client </a>
-                </div>
-                <form th:action="@{/logout}" method="post">
-                    <input type="submit" value="Sign Out"/>
-                </form>
-            </div>
-        </footer>
-    </div>
+    <div th:replace="Fragments/footer :: footer"></div>
+
+
 
 
 </body>
