+++ src/main/resources/templates/Temp/createTemp.html
@@ -9,6 +9,7 @@
     <!-- CSS + Bootstrap CSS -->
     <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
+    <script src="https://kit.fontawesome.com/610665bae1.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
 
     <title>Formulaire création température frigo</title>
@@ -17,68 +18,59 @@
 
 <body class="container">
 
-<!-- ------------------------    HEADER    ---------------------------- -->
+    <!-- ------------------------    HEADER    ---------------------------- -->
 
-<div class="row">
-
-    <header class="col-12" role="banner">
-        <img id="floatright" th:src="@{/pics/logo.png}" alt="logo logiciel" width="100" height="100">
-        <p class="gras">TPR - Traçage Produits Restaurant</p>
-        <p class="d-none d-md-none d-lg-block"> Logiciel de gestion de stock restaurant  </p>
-        <p th:inline="text">Bienvenue [[${#httpServletRequest.remoteUser}]]!</p>
+    <div th:replace="Fragments/header :: header"></div>
 
 
-    </header>
-
-</div>
+    <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
 
 
-<!--  --------------------    MAIN DE LA PAGE     -------------------  -->
-
-<div class="row">
-    <h1>Insertion Température Frigo</h1>
-
-
-    <form th:action="@{/saveTemp}"
-          th:object="${aTemp}" method="POST">
-        <div class="form-group">
-            <label>Choix du frigo</label>
-            <select th:field="*{frigoId}">
-                <option th:each="frigo : *{frigoList}"
-                        th:value="${frigo.id}"
-                        th:text="${frigo.nomFrigo}"/>
-            </select>
-            <br/>
-        </div>
-        <div class="form-group">
-            <label>Température relevée</label>
-            <input type="text" th:field="*{temperature}" />
-            <span class="error" th:if="${#fields.hasErrors('temperature')}" th:errors="*{temperature}"></span>
-            <br/>
-        </div>
-        <input type="submit" value="Submit" />
-        <input type="reset" value="Reset" />
-    </form>
-</div>
+    <div class="row">
+        <a class="btn btn-outline-secondary" th:href="@{/index}" role="button"><i class="fas fa-home"></i> Retour au Tableau de Bord</a>
+    </div>
 
 
-<!-- ------------------------    FOOTER    ---------------------------- -->
+    <br>
+    <br>
 
-<div class="row">
-    <footer class="col-12" role="contentinfo">
-        <div class="row">
-            <div class="col-3">
-                <a href="#" title="about">A propos</a>
+
+    <div class="row">
+        <h1>Insertion Température Frigo</h1>
+
+
+        <form th:action="@{/saveTemp}"
+              th:object="${aTemp}" method="POST">
+            <div class="form-group">
+                <label>Choix du frigo</label>
+                <select th:field="*{frigoId}">
+                    <option th:each="frigo : *{frigoList}"
+                            th:value="${frigo.id}"
+                            th:text="${frigo.nomFrigo}"/>
+                </select>
+                <br>
             </div>
-            <div class="col-3">
-                <a href="mailto:g.rabiller44@gmail.com" title="lien email" target="_blank">Service Client </a>
+            <div class="form-group">
+                <label>Température relevée</label>
+                <input type="text" th:field="*{temperature}" />
+                <span class="error" th:if="${#fields.hasErrors('temperature')}" th:errors="*{temperature}"></span>
+                <br>
             </div>
-            <form th:action="@{/logout}" method="post">
-                <input type="submit" value="Sign Out"/>
-            </form>
-        </div>
-    </footer>
-</div>
+            <input type="submit" value="Submit" />
+            <input type="reset" value="Reset" />
+        </form>
+    </div>
+
+
+    <br>
+    <br>
+
+
+    <!-- ------------------------    FOOTER    ---------------------------- -->
+
+    <div th:replace="Fragments/footer :: footer"></div>
+
+
 
 </body>
 </html>
