+++ src/main/resources/templates/about.html
@@ -4,51 +4,40 @@
     <meta charset="UTF-8">
     <meta name="author" content="Guillaume Rabiller" />
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <meta name="description" content="Page mise à jour statut">
+    <meta name="description" content="Page A Propos pour l'application TPR - Traçage Produits Restaurant">
 
     <!-- CSS + Bootstrap CSS -->
     <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
 
-    <title>Formulaire mise à jour statut</title>
+    <title>A Propos - Tracer restaurant</title>
 </head>
 
-<body class="container">
+<body class="container no-padding">
 
     <!-- ------------------------    HEADER    ---------------------------- -->
 
     <div th:replace="Fragments/header :: header"></div>
 
+
+
     <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
 
 
     <div class="row">
-        <h1>Mise à jour statut</h1>
-
-        <form th:action="@{/updateStatut}"
-              th:object="${aStatut}" method="POST">
-            <div class="form-group">
-                <input type="hidden" th:field="*{id}" />
-                <br/>
-            </div>
-            <div class="form-group">
-                <label>Nom du statut</label>
-                <input type="text" th:field="*{nomStatut}" />
-                <span class="error" th:if="${#fields.hasErrors('nomStatut')}">Erreur sur le nom</span>
-                <br/>
-            </div>
-
-            <input type="submit" value="Update statut" />
-            <input type="reset" value="Reset" />
-        </form>
+        <a class="btn btn-outline-secondary" th:href="@{/index}" role="button"><i class="fas fa-home"></i> Retour au Tableau de Bord</a>
     </div>
+
+
 
 
 
     <!-- ------------------------    FOOTER    ---------------------------- -->
 
     <div th:replace="Fragments/footer :: footer"></div>
+
+
