+++ src/main/resources/templates/index.html
@@ -9,47 +9,37 @@
     <!-- CSS + Bootstrap CSS -->
     <link rel="stylesheet" th:href="@{/styles/css/style.css}"/>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Open+Sans&family=Roboto&display=swap">
+    <script src="https://kit.fontawesome.com/610665bae1.js" crossorigin="anonymous"></script>
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-giJF6kkoqNQ00vy+HMDP7azOuL0xtbfIcaT9wjKHr8RbDVddVHyTfAAsrekwKmP1" crossorigin="anonymous">
 
     <title>Projet NFA036 - Tracer restaurant</title>
 </head>
 
-<body class="container no-padding">
+<body class="container">
 
     <!-- ------------------------    HEADER    ---------------------------- -->
 
-    <!-- th:replace -->
-
-    <div class="row">
-
-        <header class="col-12" role="banner">
-            <img id="floatright" th:src="@{/pics/logo.png}" alt="logo logiciel" width="100" height="100">
-            <p class="gras">TPR - Traçage Produits Restaurant</p>
-            <p class="d-none d-md-none d-lg-block"> Logiciel de gestion de stock restaurant  </p>
-            <p th:inline="text">Bienvenue [[${#httpServletRequest.remoteUser}]]!</p>
+    <div th:replace="Fragments/header :: header"></div>
 
 
-        </header>
-
-    </div>
 
     <!--  --------------------    MAIN DE LA PAGE     -------------------  -->
 
 
     <div class="row">
         <div class="col-6">
-            <a class="btn btn-outline-secondary" th:href="@{/readProduit}" role="button">Gestion des Produits</a>
+            <a class="btn btn-outline-secondary btn-lg btn-block" th:href="@{/readProduit}" role="button"><i class="fas fa-carrot"></i> Gestion des Produits</a>
         </div>
         <div class="col-6">
-            <a class="btn btn-outline-secondary" th:href="@{/readCategorie}" role="button">Gestion des Catégories produit</a>
+            <a class="btn btn-outline-secondary btn-lg btn-block" th:href="@{/readCategorie}" role="button"><i class="far fa-folder"></i> Gestion des Catégories produit</a>
         </div>
     </div>
     <div class="row">
-        <div class="col-6">
-            <a class="btn btn-outline-secondary" th:href="@{/readUtil}" role="button">Gestion des utilisateurs</a>
+        <div class="col">
+            <a class="btn btn-outline-secondary btn-lg btn-block" th:href="@{/readUtil}" role="button"><i class="fas fa-users"></i> Gestion des utilisateurs</a>
         </div>
-        <div class="d-grip gap-2 col-6">
-            <a class="btn btn-outline-secondary" th:href="@{/readFrigo}" role="button">Gestion des frigos</a>
+        <div class="col">
+            <a class="btn btn-outline-secondary btn-lg btn-block" th:href="@{/readFrigo}" role="button"><i class="fas fa-boxes"></i> Gestion des frigos</a>
         </div>
     </div>
 
@@ -68,46 +58,35 @@
                 <th>DLC</th>
                 <th>Statut</th>
             </tr>
-            <tr th:each="stock:${stockList}" th:if="${stock.statut} != 'En Stock'" th:style="${stock.colorStatut} == 2 ? 'background: orange' : 'background: red'  ">
+            <tr th:each="stock:${stockList}" th:if="${stock.statut} != 'En Stock'">
                 <td th:text="${stock.id}"></td>
                 <td th:text="${stock.nomProduit}"></td>
                 <td th:text="${stock.categorie}"></td>
                 <td th:text="${stock.dateEntree}"></td>
                 <td th:text="${stock.dlc}"></td>
-                <td th:text="${stock.statut}"></td>
+                <td th:text="${stock.statut}" th:style="${stock.colorStatut} == 2 ? 'background: #CD710B' : 'background: #A73000'  "></td>
             </tr>
         </table>
     </div>
 
-    <a class="btn btn-info" th:href="@{/readStock}" role="button">Gestion du stock</a>
+    <a class="btn btn-outline-primary" th:href="@{/readStock}" role="button"><i class="fas fa-utensils"></i> Gestion du stock</a>
 
     <br>
     <br>
     <h2>Relevé des températures frigo</h2>
 
     <div class="row">
-        <table border="1" class="table table-striped table-hover">
-            <tr>
-                <th>id</th>
-                <th>Nom du Produit</th>
-                <th>Catégorie du produit</th>
-                <th>Date / heure d'entrée</th>
-                <th>DLC</th>
-                <th>Statut</th>
-            </tr>
-            <tr th:each="stock:${stockList}" th:if="${stock.statut} != 'En Stock'" th:style="${stock.colorStatut} == 2 ? 'background: orange' : 'background: red'  ">
-                <td th:text="${stock.id}"></td>
-                <td th:text="${stock.nomProduit}"></td>
-                <td th:text="${stock.categorie}"></td>
-                <td th:text="${stock.dateEntree}"></td>
-                <td th:text="${stock.dlc}"></td>
-                <td th:text="${stock.statut}"></td>
+        <table th:each="temp:${tempList}"  border="1" class="table table-striped table-hover col">
+            <tr >
+                <td th:text="${temp.getNomFrigo()}"></td>
+                <td th:text="${temp.getTempMatin()}"></td>
+                <td th:text="${temp.getTempAprem()}"></td>
             </tr>
         </table>
     </div>
 
 
-    <a class="btn btn-info" th:href="@{/createTemp}" role="button">Ajout d'une température</a>
+    <a class="btn btn-outline-primary" th:href="@{/createTemp}" role="button"><i class="fas fa-temperature-low"></i> Ajout d'une température</a>
 
 
 
@@ -118,22 +97,7 @@
 
     <!-- ------------------------    FOOTER    ---------------------------- -->
 
-    <div class="row">
-        <footer class="col-12" role="contentinfo">
-            <div class="row">
-                <div class="col-3">
-                    <a href="#" title="about">A propos</a>
-                </div>
-                <div class="col-3">
-                    <a href="mailto:g.rabiller44@gmail.com" title="lien email" target="_blank">Service Client </a>
-                </div>
-                <form th:action="@{/logout}" method="post">
-                    <input type="submit" value="Sign Out"/>
-                </form>
-                <a  th:href="@{/readStatut}" role="button">Gestion des Statuts</a>
-            </div>
-        </footer>
-    </div>
+    <div th:replace="Fragments/footer :: footer"></div>
 
 
 </body>
